+++ src/af/xap/xp/xap_String_Id.h
@@ -341,6 +341,7 @@ dcl(TB_InsertNewTable,                   "Insert New Table")
 dcl(TB_Rows_x_Cols_Table,                "Table")
 dcl(TB_ClearBackground,"Clear Background")
 dcl(TB_ClearForeground,"Clear Foreground")
+dcl(TB_CustomColor,"Custom Color...")
 
 /*
 	  Language names, arranged alphabetically.